# Add performance benchmarks for election execution