--- a/abft/atropos_chain_test.go
+++ b/abft/atropos_chain_test.go
@@ -25,6 +25,14 @@ func TestRegressionData_SonicNetwork(t *testing.T) {
 	testRegressionData(t, "testdata/events-8000-partial.db")
 }
 
+func BenchmarkElectionFantomNetwork(b *testing.B) {
+	benchmarkElection(b, "testdata/events-5577.db")
+}
+
+func BenchmarkElectionSonicNetwork(b *testing.B) {
+	benchmarkElection(b, "testdata/events-8000-partial.db")
+}
+
 func testRegressionData(t *testing.T, dbPath string) {
 	conn, err := sql.Open("sqlite3", dbPath)
 	if err != nil {
@@ -55,28 +63,20 @@ func benchmarkElection(b *testing.B, dbPath string) {
 		b.Fatal(err)
 	}
 
-    b.ResetTimer()
+	b.ResetTimer()
 
 	for range b.N {
-	    for epoch := epochMin; epoch <= epochMax; epoch++ {
-            b.StopTimer()
-            testLachesis, eventStore, _, orderedEvents, err := SetupElection(conn, epoch);
-            if err != nil {
-                b.Fatal(err)
-            }
+		for epoch := epochMin; epoch <= epochMax; epoch++ {
+			b.StopTimer()
+			testLachesis, eventStore, _, orderedEvents, err := SetupElection(conn, epoch)
+			if err != nil {
+				b.Fatal(err)
+			}
 
-            b.StartTimer()
-            if err := ExecuteElection(testLachesis, eventStore, orderedEvents); err != nil {
-		        b.Fatal(err)
-            }
-        }
-    }
-}
-
-func BenchmarkElectionFantom(b *testing.B) {
-	benchmarkElection(b, "testdata/events-5577.db")
-}
-
-func BenchmarkElectionSonic(b *testing.B) {
-	benchmarkElection(b, "testdata/events-8000-partial.db")
+			b.StartTimer()
+			if err := ExecuteElection(testLachesis, eventStore, orderedEvents); err != nil {
+				b.Fatal(err)
+			}
+		}
+	}
 }