- Added benchmarking to time the election execution code (ignoring database querying in setup). 
- Refactored CheckEpochAgainstDB to use common code.  
- Chose to use b.Stop/StartTimer() to avoid storing multiple instances of CoreLachesis, EventStore etc. It seems to work fine and simplifies the code. b.ResetTimer is done for sanity reasons and has been recommended (in forums) if using Stop/StartTimer. It can probably be removed however with no harm. 
- Ran go fmt 

Tests continue to pass and benchmarks seem realistic: 

>psubotic@Pavles-MacBook-Pro-2 abft % go test
>PASS
>ok  	github.com/0xsoniclabs/consensus/abft	14.601s
>psubotic@Pavles-MacBook-Pro-2 abft % go test -bench .
>goos: darwin
>goarch: arm64
>pkg: github.com/0xsoniclabs/consensus/abft
>cpu: Apple M3 Pro
>BenchmarkElectionFantom-12    	       1	2670004292 ns/op
>BenchmarkElectionSonic-12     	       1	4839811750 ns/op
>PASS
>ok  	github.com/0xsoniclabs/consensus/abft	22.445s

As per @Tjofil  suggestion confirmed dbchecker builds