@psubotic 2025-03-18
> Why capitalize all the function names?

fixed.